+++ admin/placed_orders.php
@@ -14,15 +14,20 @@ if(isset($_POST['update_payment'])){
    $order_id = $_POST['order_id'];
    $payment_status = $_POST['payment_status'];
    $payment_status = filter_var($payment_status, FILTER_SANITIZE_SPECIAL_CHARS);
+   
+   if(is_numeric($order_id)){
    $update_payment = $conn->prepare("UPDATE orders SET payment_status = ? WHERE id = ?");
    $update_payment->execute([$payment_status, $order_id]);
    $message[] = 'payment status updated!';
+   }
 }
 
 if(isset($_GET['delete'])){
    $delete_id = $_GET['delete'];
+   if(is_numeric($delete_id)){
    $delete_order = $conn->prepare("DELETE FROM orders WHERE id = ?");
    $delete_order->execute([$delete_id]);
+   }
    header('location:placed_orders.php');
 }
